backend/requirements.txt
fastapi==0.111.0
starlette==0.37.2
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic[email]==2.7.1
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
openai==1.30.1
chromadb==0.5.0
numpy<2
pymupdf4llm==0.0.17
celery==5.4.0
redis==5.0.7
tenacity==8.4.1
python-docx==1.1.2
jinja2==3.1.4
weasyprint==65.1
pydyf==0.11.0
httpx==0.27.0
aiofiles==23.2.1
